+++ src/Main.hs
@@ -3,8 +3,8 @@
 
 module Main (main) where
 
-import System.Environment  (getArgs)
-import Test.Tasty.Run      (run)
+import           Test.Tasty.Prelude
+import           Test.Tasty.Run     (run)
 
 main :: IO ()
 main = getArgs >>= run
